[llvm] Move Opcode, Instruction to parent directory, add Alloca instruction
