+++ vexe-llvm/src/main/kotlin/dev/supergrecko/vexe/llvm/ir/Instruction.kt
@@ -1,19 +1,16 @@
-package dev.supergrecko.vexe.llvm.ir.instructions
+package dev.supergrecko.vexe.llvm.ir
 
 import dev.supergrecko.vexe.llvm.internal.contracts.Unreachable
 import dev.supergrecko.vexe.llvm.internal.contracts.Validatable
 import dev.supergrecko.vexe.llvm.internal.util.fromLLVMBool
 import dev.supergrecko.vexe.llvm.internal.util.wrap
-import dev.supergrecko.vexe.llvm.ir.Value
-import dev.supergrecko.vexe.llvm.ir.BasicBlock
-import dev.supergrecko.vexe.llvm.ir.MetadataEntries
 import dev.supergrecko.vexe.llvm.ir.values.DebugLocationValue
 import dev.supergrecko.vexe.llvm.ir.values.MetadataValue
 import org.bytedeco.javacpp.SizeTPointer
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
 
-public class Instruction internal constructor() : Value(),
+public open class Instruction internal constructor() : Value(),
     DebugLocationValue, Validatable {
     public override var valid = true
