+++ vexe-llvm/src/main/kotlin/dev/supergrecko/vexe/llvm/ir/values/constants/ConstantFloat.kt
@@ -2,7 +2,7 @@ package dev.supergrecko.vexe.llvm.ir.values.constants
 
 import dev.supergrecko.vexe.llvm.internal.util.fromLLVMBool
 import dev.supergrecko.vexe.llvm.ir.Value
-import dev.supergrecko.vexe.llvm.ir.instructions.RealPredicate
+import dev.supergrecko.vexe.llvm.ir.RealPredicate
 import dev.supergrecko.vexe.llvm.ir.types.FloatType
 import dev.supergrecko.vexe.llvm.ir.types.IntType
 import dev.supergrecko.vexe.llvm.ir.values.ConstantValue
